refactor: implement TestConfig
