+++ src/test/java/com/programmers/vouchermanagement/TestConfig.java
@@ -1,0 +1,40 @@
+package com.programmers.vouchermanagement;
+
+import com.zaxxer.hikari.HikariDataSource;
+import org.springframework.boot.jdbc.DataSourceBuilder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
+
+import javax.sql.DataSource;
+
+@Configuration
+@ComponentScan(
+        basePackages = {"com.programmers.vouchermanagement"}
+)
+public class TestConfig {
+    @Bean
+    public DataSource dataSource() {
+        var dataSource = DataSourceBuilder.create()
+                .url("jdbc:mysql://localhost:3306/test")
+                .username("root")
+                .password("980726")
+                .type(HikariDataSource.class)
+                .build();
+        dataSource.setMaximumPoolSize(1000);
+        dataSource.setMinimumIdle(100);
+        return dataSource;
+    }
+
+    @Bean
+    public JdbcTemplate jdbcTemplate(DataSource dataSource) {
+        return new JdbcTemplate(dataSource);
+    }
+
+    @Bean
+    public NamedParameterJdbcTemplate namedParameterJdbcTemplate(JdbcTemplate jdbcTemplate) {
+        return new NamedParameterJdbcTemplate(jdbcTemplate);
+    }
+}
