+++ src/test/java/com/programmers/vouchermanagement/customer/repository/CustomerJDBCRepositoryTest.java
@@ -1,20 +1,15 @@
 package com.programmers.vouchermanagement.customer.repository;
 
+import com.programmers.vouchermanagement.TestConfig;
 import com.programmers.vouchermanagement.customer.domain.Customer;
-import com.zaxxer.hikari.HikariDataSource;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.jdbc.DataSourceBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
+import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringJUnitConfig;
+import org.springframework.test.context.support.AnnotationConfigContextLoader;
 
-import javax.sql.DataSource;
 import java.util.List;
 import java.util.UUID;
 
@@ -22,6 +17,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringJUnitConfig
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
+@ContextConfiguration(classes = TestConfig.class, loader = AnnotationConfigContextLoader.class)
 class CustomerJDBCRepositoryTest {
     @Autowired
     CustomerJDBCRepository customerJDBCRepository;
@@ -41,34 +37,5 @@ class CustomerJDBCRepositoryTest {
     @DisplayName("🆗 고객 정보를 저장할 수 있다.")
     void save() {
         customerJDBCRepository.save(new Customer(UUID.randomUUID(), "고객4"));
-    }
-
-    @Configuration
-    @ComponentScan(
-            basePackages = {"com.programmers.vouchermanagement"}
-    )
-    static class Config {
-        @Bean
-        public DataSource dataSource() {
-            var dataSource = DataSourceBuilder.create()
-                    .url("jdbc:mysql://localhost:3306/test")
-                    .username("root")
-                    .password("980726")
-                    .type(HikariDataSource.class)
-                    .build();
-            dataSource.setMaximumPoolSize(1000);
-            dataSource.setMinimumIdle(100);
-            return dataSource;
-        }
-
-        @Bean
-        public JdbcTemplate jdbcTemplate(DataSource dataSource) {
-            return new JdbcTemplate(dataSource);
-        }
-
-        @Bean
-        public NamedParameterJdbcTemplate namedParameterJdbcTemplate(JdbcTemplate jdbcTemplate) {
-            return new NamedParameterJdbcTemplate(jdbcTemplate);
-        }
     }
 }
