+++ src/test/java/com/programmers/vouchermanagement/wallet/repository/WalletJDBCRepositoryTest.java
@@ -1,23 +1,18 @@
 package com.programmers.vouchermanagement.wallet.repository;
 
+import com.programmers.vouchermanagement.TestConfig;
 import com.programmers.vouchermanagement.customer.domain.Customer;
 import com.programmers.vouchermanagement.customer.repository.CustomerJDBCRepository;
 import com.programmers.vouchermanagement.voucher.domain.Voucher;
 import com.programmers.vouchermanagement.voucher.domain.VoucherType;
 import com.programmers.vouchermanagement.voucher.repository.VoucherJDBCRepository;
 import com.programmers.vouchermanagement.wallet.domain.Ownership;
-import com.zaxxer.hikari.HikariDataSource;
 import org.junit.jupiter.api.*;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.jdbc.DataSourceBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
+import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringJUnitConfig;
+import org.springframework.test.context.support.AnnotationConfigContextLoader;
 
-import javax.sql.DataSource;
 import java.util.Optional;
 import java.util.UUID;
 
@@ -28,6 +23,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 @SpringJUnitConfig
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
+@ContextConfiguration(classes = TestConfig.class, loader = AnnotationConfigContextLoader.class)
 class WalletJDBCRepositoryTest {
     private final static UUID NON_EXISTENT_VOUCHER_ID = UUID.randomUUID();
     private final static UUID NON_EXISTENT_CUSTOMER_ID = UUID.randomUUID();
@@ -184,34 +180,5 @@ class WalletJDBCRepositoryTest {
         voucherJDBCRepository.delete(voucher.voucherId());
 
         assertThat(walletJDBCRepository.findCustomerByVoucherId(voucher.voucherId()).isEmpty()).isTrue();
-    }
-
-    @Configuration
-    @ComponentScan(
-            basePackages = {"com.programmers.vouchermanagement"}
-    )
-    static class Config {
-        @Bean
-        public DataSource dataSource() {
-            var dataSource = DataSourceBuilder.create()
-                    .url("jdbc:mysql://localhost:3306/test")
-                    .username("root")
-                    .password("980726")
-                    .type(HikariDataSource.class)
-                    .build();
-            dataSource.setMaximumPoolSize(1000);
-            dataSource.setMinimumIdle(100);
-            return dataSource;
-        }
-
-        @Bean
-        public JdbcTemplate jdbcTemplate(DataSource dataSource) {
-            return new JdbcTemplate(dataSource);
-        }
-
-        @Bean
-        public NamedParameterJdbcTemplate namedParameterJdbcTemplate(JdbcTemplate jdbcTemplate) {
-            return new NamedParameterJdbcTemplate(jdbcTemplate);
-        }
     }
 }
